+++ CHANGELOG.md
@@ -6,3 +6,5 @@
 
 * Implemented "put a Kerbal in space for 30 days" with ContractConfigurator config
 * Implemented the "cross heliopause" contract with a custom CC parameter class
+* First contract that researches + reveals the inner radiation belt
+